Added: ChunkedFileBlock, FromExistingNxBlock & Extended IFileData Lengths to 64-bit

+++ NexusMods.Archives.Nx.Tests/Utilities/TestOutputDataProvider.cs
@@ -7,6 +7,6 @@ public class TestOutputDataProvider : IOutputDataProvider
 {
     public string RelativePath { get; init; } = null!;
     public FileEntry Entry { get; init; }
-    public IFileData GetFileData(long start, uint length) => throw new NotImplementedException();
+    public IFileData GetFileData(ulong start, ulong length) => throw new NotImplementedException();
     public void Dispose() { }
 }
