+++ NexusMods.Archives.Nx/FileProviders/FileData/ArrayFileData.cs
@@ -12,7 +12,7 @@ public sealed unsafe class ArrayFileData : IFileData
     public byte* Data { get; }
 
     /// <inheritdoc />
-    public nuint DataLength { get; }
+    public ulong DataLength { get; }
 
     private GCHandle _handle;
     private bool _disposed;
@@ -23,12 +23,12 @@ public sealed unsafe class ArrayFileData : IFileData
     /// <param name="data">The data backed up by the array.</param>
     /// <param name="start">Start offset.</param>
     /// <param name="length">Length into the data.</param>
-    public ArrayFileData(byte[] data, long start, uint length)
+    public ArrayFileData(byte[] data, ulong start, ulong length)
     {
         _handle = GCHandle.Alloc(data, GCHandleType.Pinned);
         Data = (byte*)_handle.AddrOfPinnedObject() + start;
         // ReSharper disable once RedundantCast
-        DataLength = (nuint)length;
+        DataLength = length;
     }
 
     /// <inheritdoc />
